requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
yt-dlp==2024.03.10
motor==3.3.2
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic>=2.6.1
pydantic-settings>=2.1.0